requirements.txt
astroid==2.9.3
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
py-bcrypt==0.4
pycparser==2.21
pylint==2.12.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
